Commit 2
Reviving this very dead repo, might do some more work on it at some point but probably not. I wanted to revisit machine learning and so I decided to relearn the skills that I'd used for this, but make it slightly better. My main gripe with my former design was that it could loop around, so the snake would very quickly learn to just abuse that; find a simple pattern that has no overlapping that primarily goes down and to the right. I wanted to do something slightly better, so I redid it. Now, walls kill the snake, and that means that the snake has to properly learn. To make this work, I had to use a lot less inputs (I previously had 300), so I settled on 8; the tiles directly adjacent to the head, and the 4 directions in which the food could be, up, right, down, left. By keeping all inputs as binary and having a small number, it learns very quickly and with maybe half an hour or forty minutes of training, it can consistently hit around 15-20 points per game, which I'd say is pretty good.